+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 (WIP)
+
+* Send release version info
+* Set the correct environment
+* Allow setting Sentry hooks
+
 ## 0.1.0 (2023-05-09)
 
 * Initial release